Fixed playertimes randomly reverting to negatives.
The issue was due to the fact that srcds is 32-bit and os.clock() returns a 64-bit value on 64-bit systems. Around os.clock() time of 2140.xx, the decimal precsion increases and the negative bit gets set. So the values start counting at -2140.xxxxxx back toward 0. In order to solve this problem, we take the absolute differences in time values of the clock rather than just subtracting new - old.