Clients request settings on initial load via network.
Server properly loads settings from file.
Sandbox settings moved to plugin to register under generic settings panel.
Atmos settings added as a plugin.
Temporarily added "atmos_" to the convar whitelist until I add settings for that.
Added privileges for modifying settings and sending settings to the server.
Collapsed the use of "Categories" into "SettingsTree".
Collapsed the DoChange method into evolve:SetSetting.
Commented out debug prints.
Added evolve.settingsStructure as a point of reference for validating the hierarchy of settings.
Made evolve:SetSetting client-only for now due to referencing LocalPlayer.
evolve:GetSetting() now properly evaluates defaults so that something with the default of true won't override a value of false.
evolve:SendSettings() now checks for the privilege of "Settings: Send To Server" and defaults to broadcast to all players if called as the server.
Made evolve:SaveSettings() available to the client, though this is a mistake, because von isn't available to clients.
Made net messages use a string instead of a bit because sends were being mistaken for saves.
Server automatically rebroadcasts settings once a player saves them to the server.
Registering settings will not overwrite pre-existing settings. This may be unwanted in some cases. We'll deal with that when the time comes.
Registering settings copies data instead of referencing. If you're writing a plugin remember to use evolve:GetSettings() instead of your Settings table.
Server has the option to re-load settings post-init to overwrite any changes made in server.cfg